fix(api): a link join clears any pending email invite for that address
Someone invited by email who never redeemed the invite, then arrived through the
workspace join link instead, stayed on the Admin's pending list forever. Mirror
what PUT /v1/workspace/members already does after setMembership.

Folds the api half of the test-placement move in with it, since the new case
belongs in the file that owns the feature: apps/api/test/join-link.test.ts moves
into invitations.test.ts as a second describe, and the pure "join link liveness"
block is dropped (it restated isLiveJoinLink and the TTL constant; the 410 route
test pins that contract).
